feat(contracts): add nonce for monotonically increasing delivery ordering for `HypERC4626` (#4534)
### Description

- Added a `rateUpdateNonce` in `HypERC4626Collateral` and
`previousNonce` in `HypERC4626` to ensure we only update the
exchangeRate on the synthetic asset if the update was after the last
recorded update. This is to make sure we don't update it to a stale
exchange rate which may cause losses to users using the synthetic asset.

### Drive-by changes

- `processInboundMessage` in MockMailbox` to simulate processing of
messages out of order.

### Related issues

- fixes https://github.com/chainlight-io/2024-08-hyperlane/issues/12

### Backward compatibility

Yes

### Testing

Unit test