feat: alert instead of crashing relayer when building cursors (#4811)
### Description

- we should make sure there are no `unwrap`s outside a `tokio::spawn`,
otherwise they'll take down the entire relayer. The offenders have so
far been the indexing cursors, so this PR just logs an error and sets a
new metric flag instead of panicking: `critical_error`, an `IntGauge`
type because prometheus doesn't have boolean gauges. For now this is
only used for origin chains, but I envision this as being set when a
critical error occurs in the submitter too, in the future
- before setting the flag, cursor instantiation is now retried 10 times
(to avoid short lived RPC hiccups)

### Drive-by changes
Removes usage of `Builder` derive macro in metric types, since they're
not complex enough to warrant it

### Related issues

### Backward compatibility

Yes

### Testing

the existing e2e