feat: refactor warp monitor to use WarpCore config (#4835)
### Description

- Removes the warp monitor specific config format and all the
handcrafted yamls from infra
- Massive refactor to warp monitor to accommodate this -- tried to make
it a bit easier to work work by splitting it into a few files. The
changes are generally backward compatible with some small exceptions,
like now the metrics use the `TokenStandard` instead of `TokenType` (the
former is present in the warp core config, the latter isn't, and there's
no way atm to map from the former to the latter)
- The warp monitor now takes in a warp ID and gets its config from the
registry (recall this is the registry version defined in `.registryrc`
if running in the monorepo image)
- Deploying the warp monitor now lets you do many at a time and select
the ones to deploy interactively. This makes it a bit easier to work
with the warp IDs
- The names of the warp deploys are based on the warp IDs. Because warp
IDs can change as new chains are added, this may cause issues in the
future - we'll just need to be cautious of removing the old release to
avoid accidentally having 2 of them. For now I think this is fine. The
helm release names are also capped at 53 chars so they needed to be
truncated
- Some changes to the SDK to get the necessary data here. Happy to pull
some of these into the warp monitor if people feel they're inappropriate
for the SDK (like the `getBridgedSupply`)
- Added a test to make sure any warp route IDs in infra are found in the
registry

- Corresponding PR to the registry here
https://github.com/hyperlane-xyz/hyperlane-registry/pull/370

### Drive-by changes

<!--
Are there any minor or drive-by changes also included?
-->

### Related issues

<!--
- Fixes #[issue number here]
-->

### Backward compatibility

<!--
Are these changes backward compatible? Are there any infrastructure
implications, e.g. changes that would prohibit deploying older commits
using this infra tooling?

Yes/No
-->

### Testing

<!--
What kind of testing have these changes undergone?

None/Manual/Unit Tests
-->