feat: relayer app contexts use warp route artifacts from the registry (#4845)
### Description

This PR aims to solve 2 problems: forgetting to add warp routes as app
contexts in the relayer, and deduplicating artifacts that live in infra
and the registry

- Removes a bunch of the duplicated warp route artifacts in infra, in
favor of getting them from the registry
- Now all WarpRouteIds defined in infra will be auto-populated into the
mainnet3 relayers as app contexts, with the label as the warp route ID
- This significantly increased the size of the metric app context
configuration, which was previously supplied as an env var to the
relayers. This caused the env var size limit to be hit. As a workaround,
larger configuration strings are now moved into a configmap that will be
read in as a JSON by the relayer. I intentionally kept some of the
configs as env vars, e.g. the relayChains, because this is much more
ergonomic to change on the fly and many of us have operational muscle
memory around this that I didn't want to break.
- Note one consequence of this is that merkly's ETH app is prone to
running out of ETH and causing prep queues to rise. Because of this,
I'll ignore merkly eth from our high urgency alert

### Drive-by changes

- There are a few places we were whitelisting warp routes in the gas
config that I've now removed. From what I can tell none of these are
necessary anymore as they were intended to be temporary workarounds.
- Updated the addresses for merkly from here
https://minter.merkly.com/hyperlane/docs

### Related issues

<!--
- Fixes #[issue number here]
-->

### Backward compatibility

<!--
Are these changes backward compatible? Are there any infrastructure
implications, e.g. changes that would prohibit deploying older commits
using this infra tooling?

Yes/No
-->

### Testing

<!--
What kind of testing have these changes undergone?

None/Manual/Unit Tests
-->