feat: Sealevel IGP quoting in WarpCore / token adapters (#4923)
### Description

- Implements SVM IGP quoting. Unfortunately the warp routes themselves
don't have direct getters for this, so we first need to get the expected
destination gas amount, and then do a call to the IGP
- An unfortunate consequence is that performing a "view call" on
Sealevel (i.e. a simulateTransaction call) requires you to specify a
payer that is capable of paying for the tx as if it were real (i.e. must
have funds, must be an EOA). I couldn't find any workaround other than
just requiring the payer to be passed in :(

### Drive-by changes

- Some further support for non-overhead IGP usage / some minor renaming
to more closely match the program jargon

### Related issues

<!--
- Fixes #[issue number here]
-->

### Backward compatibility

<!--
Are these changes backward compatible? Are there any infrastructure
implications, e.g. changes that would prohibit deploying older commits
using this infra tooling?

Yes/No
-->

### Testing

I tested this with an infra script that was calling the functions
against live warp routes. Not sure what the unit test practices we have
here are though, seems like we don't have any