fix(agents/sealevel): handle non configured igp account payments (#5075)
### Description

- Handles the case where igp fees are paid to a non-configured igp
account in a backward compatible way.
- Zero out payment for non-matching igp accounts so that the db stays
continuous to prevent an infinite loop in the sequencer aware parser but
at the same time, enforce payments to the right igp account was made
before processing the message.

Note: If we want to display these neutralized payments (in say, the
Explorer), we'll need to think of a more breaking solution, which isn't
necessary today.

### Drive-by changes

None

### Related issues

- fixes https://github.com/hyperlane-xyz/issues/issues/1392

### Backward compatibility

Yes

### Testing

- integration test paying to a different newly deployed account