feat: Add optional parameter for existingProxyAdmins in getRenzoEZETHWarpConfigGenerator (#5388)
### Description
This PR addresses an issue mentioned on
[discord](https://discord.com/channels/935678348330434570/1321870185984163880/1337085307073597591)
where the existing infra checker does not fully support proxyAdmin. The
error is introduced because we reused renzo's proxyAdmin.

This change allows specification of the proxyAdmin, such that if it's
not provided, will deploy a new proxyAdmin

### Backward compatibility

Yes

### Testing
Tested by running `yarn tsx ./scripts/warp-routes/export-warp-configs.ts
-e mainnet3` which resulted in:
- no diffs in production ezEth config
- removes the proxyAdmin object from EZETHSTAGE