chore: switch to using helius's high prio fee (#5430)
The helius priority gas API started returning values that were too low
at the `recommended` over the weekend. Their recommendation was to
switch to the `high` one, which seems about as high as the `low` level
reported by quicknode.

In the past, the `high` level reported significantly higher prio fees
than needed to get inclusion so we avoided it, but maybe they recently
tweaked their algorithm