fix: Make only one retrying attempt if we get non-retryable error (#5540)
### Description

Make only one retrying attempt if we get non-retryable error.

At the moment, we are doing 12 RPC calls even if all of them returned a
non-retryable error in multicast approach of EthereumFallbackProvider.
While we still would like to wait for all providers to return their
result when we do multicast, we would like to avoid retrying since it
produces a lot of requests with the same failure result. It leads to
elevated RPC failure metrics and higher number of alerts.

### Related issues

- Contributes into
https://github.com/hyperlane-xyz/hyperlane-monorepo/issues/5452

### Backward compatibility

Yes

### Testing

TBD

---------

Co-authored-by: Danil Nemirovsky <4614623+ameten@users.noreply.github.com>