fix: Scraper Sealevel to support complex transactions in recipient calculation (#5657)
### Description

Currently, Scraper has very rudimentary logic to calculate a recipient
of a transaction, namely, the recipient is identified as any non-native
program which was invoked by the transaction. Unfortunately, when
Mailbox is invoked as part of another program, transaction contains more
than one non-native programs and Mailbox is featured as inner
instruction, not as instruction on the top level of the transaction. It
makes Scraper to report error of too many non-native transactions or
incorrect recipient.

This change will allow Scraper to search for top level program which
invokes Mailbox or IGP. Usually, this top level program is a Warp Route,
which we prefer to have as recipient.

If Scraper cannot identify Mailbox or IGP, the first non-native program
will be reported as recipient.

### Drive-by changes

- Calculation of recipient is moved to separate module.

### Related issues

- Contributes into https://github.com/hyperlane-xyz/issues/issues/1432

### Backward compatibility

Yes

### Testing

Unit tests

---------

Co-authored-by: Danil Nemirovsky <4614623+ameten@users.noreply.github.com>