fix(cli): checker remoteRouters check fix (#5604)
### Description

This fixes the `warp check` command because it did not correctly check
the `remoteRouters` object when included in the config. The issue was
caused by the `normalizeConfig` function, which removes occurrences of
the `address` field from an object. This converted the remoteRouter
config to an empty object, causing the check to succeed when it
shouldn't have.


![image](https://github.com/user-attachments/assets/6f6ff782-1054-4d21-9cc6-374a7d5440c3)

A new `HypTokenRouterConfigToCheck` type has been implemented, which
renames the field to `routerAddress` for checking purposes


![image](https://github.com/user-attachments/assets/d30ad7e7-633b-484b-9b72-70b196ff1ae6)

### Drive-by changes

- None

### Related issues

### Backward compatibility

- YES

### Testing

- Manual
- E2E