feat(balances): manually reduce balance thresholds (#5731)
### Description
- PR is best reviewed in two separate commits
- Commit
https://github.com/hyperlane-xyz/hyperlane-monorepo/pull/5731/commits/e60a9e93869c8a8bdf334d401fc8d4786a00f708:
Use updated token prices and latest usage to update relayer burn and
update thresholds. This is normal operations
- Commit
https://github.com/hyperlane-xyz/hyperlane-monorepo/pull/5731/commits/502c457d13ccaa1df3d6375408db31e50bba27ac:
Manually reduces the thresholds of a set of chains, the thresholds based
off the calculated burn is more than 50% less than some of the existing
thresholds for these chains. We are reducing the thresholds as
maintaining these balances are not necessary

### Testing

Manual