feat: fix url parsing when there is no protocol provided (#5740)
### Description

When the rpc/grpc url provided does not have a protocol attached (ie.
`https://`, `ftp://`, etc), it will show up as unknown in our metrics.
This adds a fix for when `Url` lib is not able to parse it properly
because of missing protocol, we manually parse the url for our metrics.

 - I made it so, default ports also show up in the metrics
   - ie. `https://rpc.example.com/1235/1243243` -> `rpc.example.com:443`
- because we strip the `https://` in our metrics, so its hard to tell
which port its actually using
   - not sure how useful this is

### Backward compatibility

yes

### Testing

Added unit test to make sure different url formats work with metrics