feat(submitter): inclusion stage implementation (#5785)
### Description

- Finishes the Building Stage and Inclusion Stage implementations and
unit tests them to high coverage
- Adds unlimited retries to all fallible network calls. The reasoning is
that if the network is problematic, it doesn't make sense to move on to
a different payload / tx since those will have the same issue
- Changes return type of `AdaptsChain::build_transactions` to
`TxBuildingResult`, to know which payloads to drop if building a tx for
that was unsuccessful
- Refactors testing utils to reuse across all stages

### Drive-by changes

- Adds enum error types across the `submitter` crate. The point is to
eventually use these to know exactly when it's worth escalating the
price or drop the tx
- Changes the sealevel `get_estimated_costs_for_instruction` to return
an `Option`, where `None` represents failed simulation and `Some`
represents successful simulation. This helps inform whether to drop
payloads / txs
- Similarly changes the `estimate` and `simulate` methods on
`AdaptsChain`

### Related issues
- Fixes https://github.com/hyperlane-xyz/hyperlane-monorepo/issues/5766
- Fixes https://github.com/hyperlane-xyz/hyperlane-monorepo/issues/5789

### Backward compatibility

Yes

### Testing

Unit tests. See the cargo-llvm-cov report below (the columns are
`Function Coverage | Line Coverage | Region Coverage`)

![Screenshot 2025-03-31 at 02 49
48](https://github.com/user-attachments/assets/63949e36-32bb-4155-a972-2586379ef9db)