fix: Add timeout to some Cosmos gRPC requests (#5818)
### Description

Add timeout to Cosmos gRPC requests which we fully control. Some
requests are done via a wrapper methods (with BroadcastTxRequest, for
example) which do not expose raw Request, so we cannot set timeout on
them.

### Related issues

Relayer was hanging indefinitely on request for misbehaving gRPC
endpoint. This change is not a remedy, but hopefully we'll get a timeout
next time and will be able to switch to a different endpoint.

### Backward compatibility

Yes

### Testing

None

---------

Co-authored-by: Danil Nemirovsky <4614623+ameten@users.noreply.github.com>