refactor: add DerivedTokenRouterConfig (#5820)
### Description

This PR introduces a DerivedTokenRouterConfig type that extends hooks
and ISMs with the address field. It pulls the derived hook and ISM
config types into the corresponding types.js files, and adds functions
to safely extract the addresses from both (as it's not a trivial
expression even with better static types).

### Related issues

Resolves #4932 

### Testing

Unit tests and hardhat tests pass locally.