feat: add cosmos-sdk (#5825)
### Description

This PR adds a TypeScript SDK for the Hyperlane Cosmos Module. It
extends the base SDK from CosmJs and adds all the Hyperlane specific
queries and transactions to it so it can be used in the typescript/sdk
package to support the Hyperlane Cosmos Module there.

The next steps include implementing e2e tests where a local cosmos chain
with the hyperlane module will be running where the transactions from
this sdk will be submitted to. This implementation will follow in
separate PRs as discussed with Paul.

After that this new SDK as the CosmJs Provider in the Hyperlane SDK and
adding the CosmosHypCollateral and CosmosHypSynthetic token adapter.

### Drive-by changes

-

### Related issues

-

### Backward compatibility

Yes

### Testing

e2e tests will follow in separate PRs as discussed with Paul