feat(validator): public rpc flag & rpc info in `metadata_latest.json` (#5792)
### Description

<!--
What's included in this PR?
-->

This PR introduces a new flag `--allowPublicRpcs` that has to be
explicitly set to `true` if one wants to run a validator on public RPCs.

The Validator detects whether or not a RPC is public based on an extra
property per RPC Url. If the property is not set, the RPC is assumed to
be private. This way backwards compatibility is ensured and we can
incrementally rollout this feature by explicitly setting `public: true`
for each RPC in the `hyperlane-registry`

Also includes: #5793 

### Drive-by changes

<!--
Are there any minor or drive-by changes also included?
-->

### Related issues

Closes #5590
Closes #5591

<!--
- Fixes #[issue number here]
-->

### Backward compatibility
Full, until the Registry is updated to account for the new `public`
attribute
<!--
Are these changes backward compatible? Are there any infrastructure
implications, e.g. changes that would prohibit deploying older commits
using this infra tooling?

Yes/No
-->

### Testing

<!--
What kind of testing have these changes undergone?

None/Manual/Unit Tests
-->

Added unit tests to check for correct config parsing.

---------

Co-authored-by: Danil Nemirovsky <ameten@users.noreply.github.com>