feat: Add proxyAdmin.owner to ownerOverrides in Checker (#5931)
### Description
This PR adds the proxyAdmin.owner to the Checker ownerOverrides such
that it checks proxyAdmin.owner instead of the top-level owner.

This is needed for edge cases where we have proxyAdmin owner that is
different than the top-level owner. For example
`LUMIA/arbitrum-avalanche-base-bsc-ethereum-lumiaprism-optimism-polygon`
has not transferred their proxyAdmin.owner

### Backward compatibility
Yes -  only checks if proxyAdmin is explicitly supplied. 

### Testing
Manual, tested by rebasing ontop of another
[PR](https://github.com/hyperlane-xyz/hyperlane-monorepo/pull/5886)
which reads an offending config (LUMIA/arbitrum-avalanche-...) from the
registry