feat: metadata building fast path (#5983)
### Description

In most aggregation ISM setups, the messageIdMultisigIsm is one of the
submodules. This is the fastest one to build, and if the aggregation
threshold is 1, there's no need to make network calls to build metadata
for the other modules.

This PR adds this "fast path" to the aggregation ISM. If it fails, we
fall back to the previous behavior where all submodules are built
concurrently.

### Drive-by changes

Fixes the s3 sdk log silencing by matching against the log targets that
show up in gcp logs

### Backward compatibility

Yes

### Testing
e2e with a 1/2 threshold, looking up the log `Built metadata using fast
path` (28 occurrences), but it did fail 64 times