feat: revoke approvals in warp core (#5973)
### Description

This PR modifies the `ITokenAdapter` interface to add a new
`isRevokeApprovalRequired` function used in WarpCore to check if
allowance should be reset before setting a new allowance amount. This is
useful for tokens like USDT, where they require that the current
allowance is 0 when setting a new amount.
ref
https://etherscan.io/token/0xdac17f958d2ee523a2206206994597c13d831ec7#code#L205

### Drive-by changes

<!--
Are there any minor or drive-by changes also included?
-->

### Related issues

<!--
- Fixes #[issue number here]
-->

### Backward compatibility

- Yes

### Testing

- Manual testing locally with the UI