chore(metadata): clean up aggregation fast path (#6011)
### Description

- the logic for building aggregation ISM metadata via the fast path has
been cleaned up and broken into smaller functions
- previously when the fast path couldn't be used, a misleading warning
about a fast path failure was printed. (An example of that is when the
threshold was > 1). Since being unable to use the fast path is normal in
those cases, now a debug log is printed instead

Fixes
https://linear.app/hyperlane-xyz/issue/BACK-171/relayer-metadata-fast-path-change-should-be-cleaned-up-especially-the

### Backward compatibility

Yes

### Testing

E2E covers this