chore: cosmos addresses (#6022)
### Description

This PR improves the way cosmos addresses are displayed. In cosmos there
is like in Ethereum a private public key pair, but for every cosmos
chain there is a different address prefix. Example Osmosis
`osmo1sduu...`, Cosmos Hub `cosmos1j48f33..`, KYVE `kyve1jhf738...` and
so on. Right now always the address of the first chain gets displayed
which often has nothing to do with the chain the user has funds on and
wants to swap from/to. This has caused confusion in the past.

This PR addresses this issue by showing the Cosmos Hub address by
default if the user is connected with a cosmos wallet since Cosmos Hub
is still the standard in the cosmos ecosystem. Once the user selects a
cosmos origin chain that address is displayed instead.

This PR can be tested by checking out the Warp UI PR
https://github.com/hyperlane-xyz/hyperlane-warp-ui-template/pull/536

### Drive-by changes

-

### Related issues

-

### Backward compatibility

Yes

### Testing

Manual Testing with Warp UI

---------

Co-authored-by: Paul Balaji <10051819+paulbalaji@users.noreply.github.com>