fix: Add initializer for MovableCollateralRouter (#6097)
### Description
Since HypERC20Collateral is an upgradeable contract, we can't use
constructors to set storage. Here we add an initializer function to
`MovableCollateralRouter` and call it in the initializer of
`HypERC20Collateral`.

We've also changed the initializer function on `HypERC20Collateral` to
have a modifier of `reinitializer(2)` meaning that it can be called two
times. This allows us to call `initialize` again on all existing
instances of `HypERC20Collateral`.

### Drive-by changes


### Related issues



### Backward compatibility

Yes

### Testing

Unit