feat(submitter): evm-specific fields in Payload and Transaction (#6135)
### Description

The `VmSpecificTxData` enum is changed for the `Evm` to now wrap a
`TypedTransaction`. No changes to `Payload` are required to support EVM
chains, since @ameten suggested serializing everything as a vec of bytes
in the `data` field.

- One concern I have is that the serialization of `VmSpecificTxData` may
change as a result of this. Memory is allocated for Enums as: "The size
will be the size of an integer specifying which of the two variants it
is, plus the size of the largest variant, plus (sometimes) padding
between the integer and the payload."
([source](https://www.reddit.com/r/rust/comments/t6fzcv/comment/hzaxlrx/?utm_source=share&utm_medium=web3x&utm_name=web3xcss&utm_term=1&utm_content=share_button),
from one of the tokio maintainers). More details here too:
https://rust-lang.github.io/unsafe-code-guidelines/layout/enums.html#layout-of-a-data-carrying-enums-without-a-repr-annotation.
- The changes may cause the `VmSpecificTxData` stored in rocksdb for
sealevel to be interpreted differently now - but hopefully that's not a
big concern since it's just for Transactions on RC.

### Drive-by changes

Adds `ethers` as a dependency on the submitter, to gain access to
`TypedTransaction`

### Related issues

Fixes: https://github.com/hyperlane-xyz/hyperlane-monorepo/issues/6127

### Backward compatibility

This change may be breaking to the `VmSpecificTxData` format stored in
the RC db

### Testing

Covered by unit tests and E2E

---------

Co-authored-by: Danil Nemirovsky <4614623+ameten@users.noreply.github.com>