feat: fixes to SVM warp deploy tooling (#6178)
### Description

Two things:
1. Updates the commit for our fork of the spl lib that's used to include
https://github.com/hyperlane-xyz/solana-program-library/pull/4 which
fixes compilation
2. Our tooling generally requires Solana 1.18.18 to be used when running
our SVM tooling to perform the program deploy. Eclipse used to allow you
to deploy programs just fine with this version but Christopher and I
experienced this regression recently where the command just hangs
indefinitely. The “workaround” seems to be to downgrade to 1.17.31, but
this comes at the cost of not being able to use a CLI flag to set the
compute unit price when deploying a program
As a temporary fix, bc the compute unit price is originally set to 0 on
Eclipse anyways, I'm changing the tooling to omit this compute unit
price flag if it’s set to 0. But it’s still possible for us to hit
issues here because we bid up compute unit prices if there are issues
deploying the program (not a rare occurrence). If you do a deploy to
Eclipse and hit issues, I’d recommend just commenting out the compute
unit price flag at least for the eclipse leg

### Drive-by changes

<!--
Are there any minor or drive-by changes also included?
-->

### Related issues

<!--
- Fixes #[issue number here]
-->

### Backward compatibility

<!--
Are these changes backward compatible? Are there any infrastructure
implications, e.g. changes that would prohibit deploying older commits
using this infra tooling?

Yes/No
-->

### Testing

- `cargo +1.76.0 install spl-token-cli --git
https://github.com/hyperlane-xyz/solana-program-library --branch
dan/create-token-for-mint --rev e101cca --force` works
- confirmed not providing the flag works